requirements.txt
asgiref==3.5.0
django==3.2.10
django-cors-headers==3.12.0
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
pillow==9.0.0
psycopg2-binary==2.8.6
sqlparse==0.4.2
drf-extra-fields==3.4.0
